Fixed two bugs
Fixed the bug where you clicked on a point and it moved two at once, also fixed the bug where if you added multiple points on the same coordinates it would break the program. That was because the coordinates were strings. Also changed point generation, they now randomly generate. (plus changed the colors).